+++ reconstruction/CaloHitBuilder/src/helper/CaloHitCollection.cxx
@@ -70,7 +70,7 @@ bool CaloHitCollection::retrieve( TVector3 &pos , xAOD::CaloHit *&hit) const
   int phiBin = findIndex( m_phi_bins, phi );
   
   if (etaBin!=-1 && phiBin!=-1){    
-    unsigned int hash = m_detector*1e9 + m_sampling*1e7 + m_segmentation*1e6 + ( etaBin*(m_phi_bins.size()-1) + phiBin );
+    unsigned long int hash = m_detector*1e9 + m_sampling*1e7 + m_segmentation*1e6 + ( etaBin*(m_phi_bins.size()-1) + phiBin );
     if (m_collection.count(hash)>0){
       hit = m_collection.at(hash);
       return true;
